+++ src/melee/ft/chara/ftKirby/ftkirbyspecialmars.c
@@ -396,7 +396,7 @@ void ftKb_SpecialNMs_8010BC40(Fighter_GObj* gobj)
     Fighter* fp = GET_FIGHTER(gobj);
     s32 i;
 
-    if ((s32) fp->u.gw.x2238_panicCharge == 0x12) {
+    if (fp->u.gw.x2238_panicCharge == 0x12) {
         i = ftKb_MS_MsSpecialNLoop;
     } else {
         i = ftKb_MS_FeSpecialNLoop;
@@ -410,7 +410,7 @@ void ftKb_SpecialNMs_8010BC90(Fighter_GObj* gobj)
     Fighter* fp = GET_FIGHTER(gobj);
     s32 i;
 
-    if ((s32) fp->u.gw.x2238_panicCharge == 0x12) {
+    if (fp->u.gw.x2238_panicCharge == 0x12) {
         i = ftKb_MS_MsSpecialAirNLoop;
     } else {
         i = ftKb_MS_FeSpecialAirNLoop;
