+++ src/melee/ft/chara/ftKirby/ftkirbyspecialn.c
@@ -578,7 +578,7 @@ void ftKb_SpecialN_800F5BA4(Fighter* fp)
     if ((s32) fp->kind == FTKIND_KIRBY &&
         (s32) fp->u.kb.hat.kind != FTKIND_KIRBY && !fp->u.kb.hat.x8_b0 &&
         (u32) fp->victim_gobj == 0U &&
-        (u32) fp->dmg.x1860_element != HitElement_Cape &&
+        fp->dmg.x1860_element != HitElement_Cape &&
         HSD_Randi((s32) da->specialn_odds_lose_ability_on_hit) == 0)
     {
         ftKb_SpecialN_800F5D04(fp->gobj, 1);
