+++ src/melee/mn/mngallery.c
@@ -93,7 +93,7 @@ void mnGallery_80258A08(HSD_GObj* gobj, u16 width, u16 height, u32 priority)
     Vec3 interest = { 0, 0, 0 };
 
     far = 2.0f;
-    bottom = (f32) (s32) - (s32) height;
+    bottom = (f32) (-(s32) height);
     right = (f32) width;
     top = 0.0f;
     left = 0.0f;
